ボトムナブのアクティブインデックスの初期値をuseContextを使って修正
