+++ src/components/atoms/HeaderTab.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 import { useMemo, useContext, FC } from 'react';
 import clsx from 'clsx';
-import { ActiveIndexContext, Context } from '../organisms/Header';
 import type { PageTitle, PagePath } from '~/types';
+import { ActiveIndexContext, Context } from '../ContextProvider';
 
 type HeaderTabProps = {
   title: PageTitle;
