+++ src/pages/_app.tsx
@@ -4,6 +4,7 @@ import type { AppProps } from 'next/app';
 import NProgress from 'nprogress';
 import 'nprogress/nprogress.css';
 import { Layout } from '../Layout';
+import { ContextProvider } from '~/components/ContextProvider';
 
 NProgress.configure({ showSpinner: false, speed: 270, minimum: 0.08 });
 
@@ -17,9 +18,11 @@ function MyApp({ Component, pageProps }: AppProps) {
   });
 
   return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
+    <ContextProvider>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </ContextProvider>
   );
 }
 export default MyApp;
